+++ bcbio/broad/__init__.py
@@ -49,8 +49,8 @@ def _get_gatk_opts(config, names, tmp_dir=None, memscale=None, include_gatk=True
     jvm_opts += get_default_jvm_opts(tmp_dir)
     return jvm_opts + opts
 
-def get_gatk_framework_opts(config, tmp_dir=None, memscale=None):
-    return _get_gatk_opts(config, ["gatk-framework", "gatk"], tmp_dir, memscale)
+def get_gatk_framework_opts(config, tmp_dir=None, memscale=None, include_gatk=True):
+    return _get_gatk_opts(config, ["gatk-framework", "gatk"], tmp_dir, memscale, include_gatk=include_gatk)
 
 def get_gatk_opts(config, tmp_dir=None, memscale=None, include_gatk=True):
     return _get_gatk_opts(config, ["gatk", "gatk-framework"], tmp_dir, memscale,